Moved /delete commands to a new feature.
