+++ features/delete/__init__.py
@@ -1,0 +1,2 @@
+from . import command
+from .feature import *
